refactor: рефакторит код проекта и изменяет файловую структуру
