+++ src/apps/cli/app/lib/command-parser/command-parser.interface.ts
