+++ src/apps/cli/app/lib/command-parser/parsed-commands.type.ts
