+++ src/apps/cli/command/command.interface.ts
