+++ src/apps/cli/command/generate.command.ts
