+++ src/apps/cli/command/lib/tsv/offer-generator/index.ts
