+++ src/apps/cli/command/lib/tsv/offer-generator/tsv-offer-generator.interface.ts
