+++ src/apps/cli/command/version.command.ts
