+++ src/apps/cli/mock/index.ts
@@ -1,0 +1,1 @@
+export type { TMockServerData } from './mock-server-data.type.js';
