+++ src/apps/cli/mock/mock-server-data.type.ts
