+++ src/apps/rest/config/index.ts
@@ -1,0 +1,2 @@
+export * from './interface/index.js';
+export { RESTConfig } from './rest.config.js';
