+++ src/apps/rest/config/interface/index.ts
@@ -1,0 +1,2 @@
+export type { IRESTConfig } from './rest-config.interface.js';
+export type { IRESTSchema } from './rest-schema.interface.js';
