+++ src/apps/rest/config/interface/rest-config.interface.ts
