+++ src/apps/rest/config/interface/rest-schema.interface.ts
@@ -1,0 +1,5 @@
+export interface IRESTSchema {
+  PORT: number;
+  SALT: string;
+  DB_HOST: string;
+}
