+++ src/apps/rest/config/rest.config.ts
@@ -2,8 +2,8 @@ import { inject, injectable } from 'inversify';
 import { config } from 'dotenv';
 import { ILogger } from '@/shared/lib/index.js';
 import { Interface } from '../const/index.js';
-import { IRESTConfig } from './rest-config.interface.js';
-import { IRESTSchema, restSchema } from './rest.schema.js';
+import { IRESTConfig, IRESTSchema } from './interface/index.js';
+import { restSchema } from './rest.schema.js';
 
 @injectable()
 export class RESTConfig implements IRESTConfig<IRESTSchema> {
