+++ src/apps/rest/const/interface.enum.ts
