--- src/cli/mock/index.ts
@@ -1,1 +1,0 @@
-export { TMockServerData } from './mock-server-data.type.js';
