--- src/rest/config/index.ts
@@ -1,3 +1,0 @@
-export type { IRESTConfig } from './rest-config.interface.js';
-export type { IRESTSchema } from './rest.schema.js';
-export { RESTConfig } from './rest.config.js';
