+++ src/shared/types/index.ts
@@ -1,7 +1,7 @@
-export { TComment } from './comment.type.js';
-export { TFacilities } from './facilities.type.js';
-export { THousingType } from './housing-type.type.js';
-export { TLocation } from './location.type.js';
-export { TOffer } from './offer.type.js';
-export { TUserType } from './user-type.type.js';
-export { TUser } from './user.type.js';
+export type { TComment } from './comment.type.js';
+export type { TFacilities } from './facilities.type.js';
+export type { THousingType } from './housing-type.type.js';
+export type { TLocation } from './location.type.js';
+export type { TOffer } from './offer.type.js';
+export type { TUserType } from './user-type.type.js';
+export type { TUser } from './user.type.js';
